Refactor notifications into provider modules
